Crate calimero_storage

Storage interface providing functionality for interacting with stored data.

This module provides the primary interface for interacting with the storage system, as a layer on top of the underlying database store.

Modules

address

Addressing of elements in the storage system.

collections

High-level data structures for storage.

entities

Entities for the storage system.

env

Environment bindings for the storage crate.

exports

Re-exported types, mostly for use in macros (for convenience).

index

Indexing system for efficient tree navigation.

integration

Types used for integration with the runtime.

interface

Interface for the storage system.

store

Storage operations.

sync

Synchronisation utilities for external runtimes.

Derive Macros

AtomicUnit

Derives the AtomicUnit trait for a struct.

Collection

Derives the Collection trait for a struct.