calimero_storage::address
Struct Path

pub struct Path { /* private fields */ }

Path to an Element.

Elements are stored in a hierarchical structure, and their path represents their location within that structure. Path segments are separated by a double-colon ::, and all paths are absolute, and should start with a leading separator to enforce the clarity of this plus allow for future expansion of functionality.

Paths are case-sensitive, support Unicode, and are limited to 255 characters in length. Note, the separators do NOT count towards this limit.

Paths are not allowed to be empty.

There is no formal limit to the levels of hierarchy allowed, but in practice this is limited to 255 levels (assuming a single byte per segment name).

Implementations

impl Path

pub fn new<S: AsRef<str>>(path: S) -> Result<Self, PathError>

Creates a new Path from a string.

Parameters

path - The path to the Element.

Errors

An error will be returned if:

The path is empty, including if it contains only separators.

The path is too long (the maximum length allowed is 255 characters).

Any of the path segments are empty.

The path is not absolute, i.e. it does not start with a double colon separator ::.

Examples

use calimero_storage::address::Path;
Path::new("::root::node::leaf").unwrap();

pub fn depth(&self) -> usize

The number of segments in the Path.

Returns the depth of the path, which is one less than the number of segments, because the roots are level 0.

Examples

use calimero_storage::address::Path;
let path = Path::new("::root::node::leaf").unwrap();
assert_eq!(path.depth(), 2);

pub fn first(&self) -> &str

The first segment of the Path.

Returns the first segment of the path, which is the top-most in the hierarchy expressed by the path.

Examples

use calimero_storage::address::Path;
let path = Path::new("::root::node::leaf").unwrap();
assert_eq!(path.first(), "root");

pub fn is_ancestor_of(&self, other: &Self) -> bool

Checks if the Path is an ancestor of another Path.

Returns true if the Path is an ancestor of the other Path, and false otherwise. In order to be counted as an ancestor, the path must be strictly shorter than the other path, and all segments must match.

Parameters

other - The other Path to check against.

Examples

use calimero_storage::address::Path;
let path1 = Path::new("::root::node").unwrap();
let path2 = Path::new("::root::node::leaf").unwrap();
assert!(path1.is_ancestor_of(&path2));

pub fn is_descendant_of(&self, other: &Self) -> bool

Checks if the Path is a descendant of another Path.

Returns true if the Path is a descendant of the other Path, and false otherwise. In order to be counted as a descendant, the path must be strictly longer than the other path, and all segments must match.

Parameters

other - The other Path to check against.

Examples

use calimero_storage::address::Path;
let path1 = Path::new("::root::node::leaf").unwrap();
let path2 = Path::new("::root::node").unwrap();
assert!(path1.is_descendant_of(&path2));

pub fn is_root(&self) -> bool

Checks if the Path is the root.

Returns true if the Path is the root, and false otherwise.

Examples

use calimero_storage::address::Path;
assert!(Path::new("::root").unwrap().is_root());
assert!(!Path::new("::root::node").unwrap().is_root());

pub fn join(&self, other: &Self) -> Result<Self, PathError>

Joins two Paths.

Joins the Path with another Path, returning a new Path that is the concatenation of the two.

Parameters

other - The other Path to join with.

Errors

An error will be returned if the resulting path would be too long.

Examples

use calimero_storage::address::Path;
let path1 = Path::new("::root::node").unwrap();
let path2 = Path::new("::leaf").unwrap();
let joined = path1.join(&path2).unwrap();
assert_eq!(joined.to_string(), "::root::node::leaf");

pub fn last(&self) -> &str

The last segment of the Path.

Returns the last segment of the path, which is the bottom-most in the hierarchy expressed by the path.

Examples

use calimero_storage::address::Path;
let path = Path::new("::root::node::leaf").unwrap();
assert_eq!(path.last(), "leaf");

pub fn parent(&self) -> Option<Self>

The parent of the Path.

Returns the parent of the Path, which is the path with the last segment removed. If the path is already at the root, then None is returned.

Examples

use calimero_storage::address::Path;
let path = Path::new("::root::node::leaf").unwrap();
assert_eq!(path.parent().unwrap().to_string(), "::root::node");
assert_eq!(Path::new("::root").unwrap().parent(), None);

pub fn segment(&self, index: usize) -> Option<&str>

The segment at a given index.

Returns the segment at the given index, or None if the index is out of bounds. Note that the root is at index 0.

Parameters

index - The index of the segment to retrieve.

Examples

use calimero_storage::address::Path;
let path = Path::new("::root::node::leaf").unwrap();
assert_eq!(path.segment(1).unwrap(), "node");
assert_eq!(path.segment(3), None);

pub fn segments(&self) -> impl Iterator<Item = &str>

The segments of the Path.

Returns the segments of the path as a vector of strings.

Examples

use calimero_storage::address::Path;
let path = Path::new("::root::node::leaf").unwrap();
assert_eq!(path.segments().collect::<Vec<_>>(), vec!["root", "node", "leaf"]);