calimero_storage::collections::unordered_set
Struct UnorderedSet

pub struct UnorderedSet<V, S: StorageAdaptor = MainStorage> { /* private fields */ }

A set collection that stores unqiue values once.

Implementations

impl<V> UnorderedSet<V, MainStorage>
where V: BorshSerialize + BorshDeserialize,

pub fn new() -> Self

Create a new set collection.

impl<V, S> UnorderedSet<V, S>
where V: BorshSerialize + BorshDeserialize, S: StorageAdaptor,

pub fn insert(&mut self, value: V) -> Result<bool, StoreError>
where V: AsRef<[u8]> + PartialEq,

Insert a value pair into the set collection if the element does not already exist.

Errors

If an error occurs when interacting with the storage system, or a child Element cannot be found, an error will be returned.

pub fn iter(&self) -> Result<impl Iterator<Item = V> + '_, StoreError>

Get an iterator over the items in the set.

Errors

If an error occurs when interacting with the storage system, or a child Element cannot be found, an error will be returned.

pub fn len(&self) -> Result<usize, StoreError>

Get the number of items in the set.

Errors

If an error occurs when interacting with the storage system, or a child Element cannot be found, an error will be returned.

pub fn contains<Q>(&self, value: &Q) -> Result<bool, StoreError>
where V: Borrow<Q>, Q: PartialEq + ?Sized + AsRef<[u8]>,

Get the value for a key in the set.

Errors

If an error occurs when interacting with the storage system, or a child Element cannot be found, an error will be returned.

pub fn remove<Q>(&mut self, value: &Q) -> Result<bool, StoreError>
where V: Borrow<Q>, Q: PartialEq + AsRef<[u8]> + ?Sized,

Remove a key from the set, returning the value at the key if it previously existed.

Errors

If an error occurs when interacting with the storage system, or a child Element cannot be found, an error will be returned.

pub fn clear(&mut self) -> Result<(), StoreError>

Clear the set, removing all items.

Errors

If an error occurs when interacting with the storage system, or a child Element cannot be found, an error will be returned.