calimero_storage
Module entities

Entities for the storage system.

This module contains the entities that are used to represent the elements in the storage system. These are the building blocks that are used to construct the hierarchy of the storage, and to apply updates to the elements.

Design: Elements, data, and atomic units

There are a number of requirements that need to be catered for by the solution, and it is worth exploring some of the possible structures and approaches that have been considered.

Considerations

Let’s examine first principles:

Elements are saved, and Elements are structs, with system metadata.

Elements also contain user data, represented by Data.

User types need to have an easy interface, using an AtomicUnit annotation.

It is possible that Element should be a trait, and that the AtomicUnit trait should be a superset of Element, and that the AtomicUnit proc macro would then apply both Element and AtomicUnit. However, this would still need a Data struct to be constructed, owned by the Element.

Therefore, if we say that Element remains a struct, for simplicity, then what exactly is being annotated? If the user creates a Person, is that Person an Element or an AtomicUnit?

It seems that the Person must be an AtomicUnit, which means it cannot be Data with this structure, as it would need to contain Data. Now, it initially seemed that AtomicUnit should extend Data. But if AtomicUnit is the Data, then the Person annotation would be misleading if it was applying to the internal Data and not to the Person. This would not be idiomatic nor a good pattern to follow.

So… if the Person is in fact the AtomicUnit, and it has data fields, where do we get Data from now? We can’t really have Data being an Element, so Data has to be owned by Element, and where does the Element come from?

The relationship between these entities is critical to get right. As much as possible of the internals should be abstracted away from the user, with the presented interface being as simple as possible.

Option One

Person is an AtomicUnit.

Person defines fields — some marked as private.

To satisfy AtomicUnit, Person has a storage field added.

Person.storage is an Element, and contains a PersonData struct.

PersonData is constructed by the macro, and contains a copy of the fields from Person that are not private. These fields could potentially be borrowed, referencing the fields on the Element.

It’s then unclear if Person (i.e. AtomicUnit) needs a save() method, or whether it should be passed to Interface.

This suffers from a number of issues, including:

The Person is now a leaky abstraction, as it exposes the internals of the Element and PersonData.

The Person is now a pain to work with, as it has to be constructed in a specific way, including the construction of PersonData.

The very existence of PersonData is icky.

Option Two

Person is an AtomicUnit.

Person defines fields — some marked as private.

Element becomes a trait instead of a struct.

AtomicUnit demands Element.

Person is now therefore an Element.

Person.data contains a PersonData struct.

PersonData is constructed by the macro, and contains a copy of the fields from Person that are not private. Again, these fields could potentially be borrowed, referencing the fields on the Element.

Person does not appear to need a save() now, as Interface.save() accepts Element, and therefore Person.

This is also problematic:

The Element has a lot on it, which clutters up the Person.

It exposes the internals, which is undesirable.

Option Three

Person is an AtomicUnit.

Person defines fields — some marked as private.

Element remains a struct.

Interface accepts Data not Element.

AtomicUnit extends Data.

Person is therefore Data.

Person has a storage field added, to hold an Element.

A circular reference is carefully created between Person and Element, with Element holding a weak reference to Person, and Person having a strong reference to (or ownership of) Element.

When accepting Data, i.e. a Person here, the Interface can call through Person.storage to get to the Element, and the Element can access the Data through the Weak.

The macro does not need to create another struct, and the saving logic can skip the fields marked as private.

Person does not appear to need a save(), as Interface.save() does not need to be changed.

The disadvantages of this approach are:

The circular reference is a bit of a pain, and could be a source of bugs.

In order to achieve the circular reference, the Element would need to wrap the Data in a Weak, which means that all use of the Data would need to be through the reference-counted pointer, which implies wrapping Data in an Arc upon construction.

In order to gain mutability, some kind of mutex or read-write lock would be required. Imposing a predetermined mutex type would be undesirable, as the user may have their own requirements. It’s also an annoying imposition.

Due to the Data (i.e. Person) being wrapped in an Arc, Default cannot be implemented.

Option Four

Person is an AtomicUnit.

Person defines fields — some marked as private.

Element remains a struct.

Interface accepts Data not Element.

AtomicUnit extends Data.

Person is therefore Data.

Person has a storage field added, to hold (and own) an Element.

When accepting Data, i.e. a Person here, the Interface can call through Person.storage to get to the Element.

The macro does not need to create another struct, and the saving logic can skip the fields marked as private.

Person does not appear to need a save(), as Interface.save() does not need to be changed.

This is essentially the same as Option Three, but without the circular reference. The disadvantages are:

There is no way for the Element to access the Data directly. This forces a different approach to those operations that require Data.

The Element’s operations can accept a Data instead of looking it up directly, but this loses structural assurance of the relationship.

Passing the Data to the Element’s operations does work around the first issue, albeit in a slightly unwieldy way. The second issue can be mitigated by performing a check that the Element owned by the passed-in Data is the same instance as the Element that the operation is being called on. Although this is a runtime logic check, it is a simple one, and can be entirely contained within the internal logic, which remains unexposed.

Option Five

For the sake of completeness, it’s worth noting that technically another option would be to use generics, such as Element<D: Data>, but this leads to unnecessary complexity at this stage (including the potential to have to use phantom data), with no tangible benefits. Therefore this has not been discussed in detail here, although it has been carefully considered.

Conclusion

Option Four is the approach chosen, as it balances all of the following aspirations:

Simple to use.

Allows full ownership and mutability of the user’s types.

Abstracts all storage functionality to the storage internals.

Achieves reliability.

Allows implementation of Default and similar without unexpected constraints.

Does not clutter the user’s types.

Does not expose the internals of the storage system.

Does not require a circular reference.

Does not impose a mutex type.

Does not force a predetermined mutex type.

Does not require a separate struct to be constructed.

Does not require references to, or clones of, saved data.

Not having a reference back to the Data from the Element is a slight trade-off, but it is a small one, sacrificing a little structural assurance and direct access for a simpler design with fewer impositions. The internal validation check is simple, and it is best to promote the elegance of the user interface over that of the internals — if there are a couple of hoops to jump through then it is best for these to be in the library code.

Design: Collections

There are three main ways to implement the collection functionality, i.e. where a parent Element has children. These are:

Struct-based: Annotate the struct as being a Collection, meaning it can then have children. In this situation the child type is supplied as an associated type. This is the most straightforward approach, but it does not allow for the parent to have multiple types of children.

Enum-based: E.g. enum ChildType { Page(Page), Author(Author) }. This is more flexible, but it requires a match to determine the child type, and although the type formality is good in some ways, it adds a level of complexity and maintenance that is not desirable.

Field-based: E.g. entity.pages and entity.authors annotated as Collection, with some way to look up which fields are needed. This is the most flexible, and the easiest developer interface to use.

The approach taken is Option 3, for the reasons given above.

Structs

ChildInfo

Summary information for the child of an Element in the storage.

Element

Represents an Element in the storage.

Metadata

The metadata for an Element.

UpdatedAt

The timestamp when the Element was last updated.

Traits

AtomicUnit

Represents an atomic unit in the storage system.

Collection

A collection of child elements in the storage system.

Data

The primary data for the Element.