calimero_storage::entities
Trait Collection

pub trait Collection {
    type Child: Data;

    // Required method
    fn name(&self) -> &str;
}

A collection of child elements in the storage system.

Collections are logical groupings of child Elements. They do not have their own storage or Element, but instead provide a way to logically group and access child elements of a specific type.

Examples

use borsh::{BorshSerialize, BorshDeserialize};
use calimero_storage_macros::{AtomicUnit, Collection};
use calimero_storage::entities::{ChildInfo, Data, Element};

#[derive(AtomicUnit, Clone, Debug, Eq, PartialEq, PartialOrd, BorshSerialize, BorshDeserialize)]
#[type_id(42)]
struct Book {
    title: String,
    pages: Pages,
    #[storage]
    storage: Element,
}

#[derive(Collection, Clone, Debug, Eq, PartialEq, PartialOrd)]
#[children(Page)]
struct Pages;

#[derive(AtomicUnit, Clone, Debug, Eq, PartialEq, PartialOrd, BorshSerialize, BorshDeserialize)]
#[type_id(43)]
struct Page {
    content: String,
    #[storage]
    storage: Element,
}

Required Associated Types

type Child: Data

The associated type of any children that the Collection may have.

Required Methods

fn name(&self) -> &str

The name of this Collection.

This is used to identify the collection when updating the index.