calimero_storage::entities
Trait Data

pub trait Data: BorshDeserialize + BorshSerialize {
    // Required methods
    fn collections(&self) -> BTreeMap<String, Vec<ChildInfo>>;
    fn element(&self) -> &Element;
    fn element_mut(&mut self) -> &mut Element;

    // Provided methods
    fn id(&self) -> Id { ... }
    fn path(&self) -> Path { ... }
}

The primary data for the Element.

This is the primary data for the Element, that is, the data that the consumer application has stored in the Element. This is the data that the Element is primarily concerned with, and for the management of which the Element exists.

This trait represents a common denominator for the various specific traits such as AtomicUnit that are used to denote exactly what the Element is representing.

Pass-through methods

The Data trait contains a number of pass-through methods that are implemented for convenience, addressing the inner Element. Potentially most of the Element methods could be added, or even Deref implemented, but for now only the most useful and least likely to be contentious methods are included, to keep the interface simple and focused.

Required Methods

fn collections(&self) -> BTreeMap<String, Vec<ChildInfo>>

Information about the Collections present in the Data.

This method allows details about the subtree structure and children to be obtained. It does not return the actual Collection types, but provides their names and child information.

fn element(&self) -> &Element

The associated Element.

The Element contains additional metadata and storage-related identification and other information for the primary data, keeping it abstracted and separate from the primary data itself.

See also

element_mut()

fn element_mut(&mut self) -> &mut Element

The associated Element, with mutability.

This function is used to obtain a mutable reference to the Element that contains the primary data.

See also

element()

Provided Methods

fn id(&self) -> Id

The unique identifier for the Element.

This is a convenience function that passes through to Element::id(). See that method for more information.

See also

Element::id()

fn path(&self) -> Path

The path to the Element in the hierarchy.

This is a convenience function that passes through to Element::path(). See that method for more information.

See also

Element::path()

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.