calimero_storage
Module interface

Interface for the storage system.

This module contains the interface for the storage system, which provides the basics of loading and saving data, but presents a number of helper methods and additional functionality to abstract away common operations.

This follows the repository pattern, where the interface is the primary means of interacting with the storage system, rather than the ActiveRecord pattern where the model is the primary means of interaction.

Synchronisation

There are two main mechanisms involved in synchronisation:

Direct changes: When a change is made locally, the resulting actions need to be propagated to other nodes.

Comparison: When a comparison is made between two nodes, the resulting actions need to be taken to bring the nodes into sync.

The entry points for synchronisation are therefore either the apply_action() method, to carry out actions; or the compare_trees() method, to compare two nodes, which will emit actions to pass to apply_action() on either the local or remote node, or both.

CRDT model

Calimero primarily uses operation-based CRDTs, also called commutative replicated data types (CmRDTs). This means that the order of operations does not matter, and the outcome will be the same regardless of the order in which the operations are applied.

It is worth noting that the orthodox CmRDT model does not feature or require a comparison activity, as there is an operating assumption that all updates have been carried out fully and reliably.

The alternative CRDT approach is state-based, also called convergent replicated data types (CvRDTs). This is a comparison-based approach, but the downside is that this model requires the full state to be transmitted between nodes, which can be costly. Although this approach is easier to implement, and fits well with gossip protocols, it is not as efficient as the operation-based model.

The standard choice therefore comes down to:

Use CmRDTs and guarantee that all updates are carried out fully and reliably, are not dropped or duplicated, and are replayed in causal order.

Use CvRDTs and accept the additional bandwidth cost of transmitting the full state for every single CRDT.

It does not fit the Calimero objectives to transmit the entire state for every update, but there is also no guarantee that all updates will be carried out fully and reliably. Therefore, Calimero uses a hybrid approach that represents the best of both worlds.

In the first instance, operations are emitted (in the form of Actions) whenever a change is made. These operations are then propagated to other nodes, where they are executed. This is the CmRDT model.

However, there are cases where a node may have missed an update, for instance, if it was offline. In this case, the node will be out of sync with the rest of the network. To bring the node back into sync, a comparison is made between the local node and a remote node, and the resulting actions are executed. This is the CvRDT model.

The storage system maintains a set of Merkle hashes, which are stored against each element, and represent the state of the element and its children. The Merkle hash for an element can therefore be used to trivially determine whether an element or any of its descendants have changed, without actually needing to compare every entity in the tree.

Therefore, when a comparison is carried out it is not a full state comparison, but a comparison of the immediate data and metadata of given element(s). This is sufficient to determine whether the nodes are in sync, and to generate the actions needed to bring them into sync. If there is any deviation detected, the comparison will recurse into the children of the element(s) in question, and generate further actions as necessary — but this will only ever examine those descendent entities for which the Merkle hash differs.

We can therefore summarise this position as being: Calimero uses a CmRDT model for direct changes, and a CvRDT model for comparison as a fallback mechanism to bring nodes back into sync when needed.

Direct changes

When a change is made locally, the resulting actions need to be propagated to other nodes. An action list will be generated, which can be made up of Add, Delete, and Update actions. These actions are then propagated to all the other nodes in the network, where they are executed using the apply_action() method.

This is a straightforward process, as the actions are known and are fully self-contained without any wider impact. Order does not strictly matter, as the actions are commutative, and the outcome will be the same regardless of the order in which they are applied. Any conflicts are handled using the last-write-wins strategy.

There are certain cases where a mis-ordering of action, which is essentially the same as having missing actions, can result in an invalid state. For instance, if a child is added before the parent, the parent will not exist and the child will be orphaned. In this situation we can either ignore the child, or we can block its addition until the parent has been added, or we can store it as an orphaned entity to be resolved later. At present we follow the last approach, as it aligns well with the use of comparisons to bring nodes back into sync. We therefore know that the node will eventually become consistent, which is all we guarantee.

TODO: Examine whether this is the right approach, or whether we should for TODO: instance block and do a comparison on the parent to ensure that local TODO: state is as consistent as possible.

Providing all generated actions are carried out, all nodes will eventually be in sync, without any need for comparisons, transmission of full states, or a transaction model (which requires linear history, and therefore becomes mathematically unsuitable for large distributed systems).

Comparison

There are a number of situations under which a comparison may be needed:

A node has missed an update, and needs to be brought back into sync (i.e. there is a gap in the instruction set).

A node has been offline, and needs to be brought back into sync (i.e. all instructions since a certain point have been missed).

A discrepancy has been detected between two nodes, and they need to be brought back into sync.

A comparison is primarily triggered from a catch-up as a proactive measure, i.e. without knowing if any changes exist, but can also arise at any point if a discrepancy is detected.

When performing a comparison, the data we have is the result of the entity being serialised by the remote note, passed to us, and deserialised, so it should be comparable in structure to having loaded it from the local database.

We therefore have access to the data and metadata, which includes the immediate fields of the entity (i.e. the AtomicUnit) and also a list of the children.

The stored list of children contains their Merkle hashes, thereby allowing us to check all children in one operation instead of needing to load each child in turn, as that would require remote data for each child, and that would not be as efficient.

If a child exists on both sides and the hash is different, we recurse into a comparison for that child.

If a child is missing on one side then we can go with the side that has the latest parent and add or remove the child.

Notably, in the case of there being a missing child, the resolution mechanism does expose a small risk of erroneous outcome. For instance, if the local node has had a child added, and has been offline, and the parent has been updated remotely — in this situation, in the absence of any other activity, a comparison (e.g. a catch-up when coming back online) would result in losing the child, as the remote node would not have the child in its list of children. This edge case should usually be handled by the specific add and remove events generated at the time of the original activity, which should get propagated independently of a sync. However, there are three extended options that can be implemented:

Upon catch-up, any local list of actions can be held and replayed locally after synchronisation. Alone, this would not correct the situation, due to last-write-wins rules, but additional logic could be considered for this particular situation.

During or after performing a comparison, all local children for an entity can be loaded by path and checked against the parent’s list of children. If there are any deviations then appropriate action can be taken. This does not fully cater for the edge case, and again would not correct the situation on its own, but additional logic could be added.

A special kind of “deleted” element could be added to the system, which would store metadata for checking. This would be beneficial as it would allow differentiation between a missing child and a deleted child, which is the main problem exposed by the edge case. However, although this represents a full solution from a data mechanism perspective, it would not be desirable to store deleted entries permanently. It would be best combined with a cut-off constraint, that would limit the time period in which a catch-up can be performed, and after which the deleted entities would be purged. This does add complexity not only in the effect on the wider system of implementing that constraint, but also in the need for garbage collection to remove the deleted entities. This would likely be better conducted the next time the parent entity is updated, but there are a number of factors to consider here.

Another way to potentially handle situations of this nature is to combine multiple granular updates into an atomic group operation that ensures that all updates are applied together. However, this remains to be explored, as it may not fit with the wider system design.

Due to the potential complexity, this edge case is not currently mitigated, but will be the focus of future development.

TODO: Assess the best approach for handling this edge case in a way that TODO: fits with the wider system design, and add extended tests for it.

The outcome of a comparison is that the calling code receives a list of actions, which can be Add, Delete, Update, and Compare. The first three are the same as propagating the consequences of direct changes, but the comparison action is a special case that arises when a child entity is found to differ between the two nodes, whereupon the comparison process needs to recursively descend into the parts of the subtree found to differ.

The calling code is then responsible for going away and obtaining the information necessary to carry out the next comparison action if there is one, as well as relaying the generated action list.

Structs

ComparisonData

Data that is used for comparison between two nodes.

Interface

The primary interface for the storage system.

Enums

Action

Actions to be taken during synchronisation.

StorageError

Errors that can occur when working with the storage system.

Type Aliases

MainInterface

Convenient type alias for the main storage system.