calimero_storage::interface
Struct Interface

#[non_exhaustive]
pub struct Interface<S: StorageAdaptor = MainStorage>(/* private fields */);

The primary interface for the storage system.

Implementations

impl<S: StorageAdaptor> Interface<S>

pub fn add_child_to<C: Collection, D: Data>( parent_id: Id, collection: &C, child: &mut D, ) -> Result<bool, StorageError>

Adds a child to a collection.

Parameters

parent_id - The ID of the parent entity that owns the Collection.

collection - The Collection to which the child should be added.

child - The child entity to add.

Errors

If an error occurs when interacting with the storage system, an error will be returned.

pub fn apply_action(action: Action) -> Result<(), StorageError>

Applies an Action to the storage system.

This function accepts a single incoming Action and applies it to the storage system. The action is deserialised into the specified type, and then applied as appropriate.

Note: In order to call this function, the caller needs to know the type of the entity that the action is for, and this type must implement the Data trait. One of the possible ways to achieve this is to use the path information externally to match against the entity type, using knowledge available in the system. It is also possible to extend this function to deal with the type being indicated in the serialised data, if appropriate, or in the ID or accompanying metadata.

After applying the Action, the ancestor hashes will be recalculated, and this function will compare them against the expected hashes. If any of the hashes do not match, the ID of the first entity with a mismatched hash will be returned — i.e. the nearest ancestor.

Parameters

action - The Action to apply to the storage system.

Errors

If there is an error when deserialising into the specified type, or when applying the Action, an error will be returned.

pub fn children_of<C: Collection>( parent_id: Id, collection: &C, ) -> Result<Vec<C::Child>, StorageError>

The children of the Collection.

This gets the children of the Collection, which are the Elements that are directly below the Collection’s owner in the hierarchy. This is a simple method that returns the children as a list, and does not provide any filtering or ordering.

Notably, there is no real concept of ordering in the storage system, as the records are not ordered in any way. They are simply stored in the hierarchy, and so the order of the children is not guaranteed. Any required ordering must be done as required upon retrieval.

Determinism

TODO: Update when the child_info field is replaced with an index.

Depending on the source, simply looping through the children may be non-deterministic. At present we are using a Vec, which is deterministic, but this is a temporary measure, and the order of children under a given path is not enforced, and therefore non-deterministic.

When the child_info field is replaced with an index, the order may be enforced using created_at timestamp, which then allows performance optimisations with sharding and other techniques.

Performance

TODO: Update when the child_info field is replaced with an index.

Looping through children and combining their hashes into the parent is logically correct. However, it does necessitate loading all the children to get their hashes every time there is an update. The efficiency of this can and will be improved in future.

Parameters

parent_id - The ID of the parent entity that owns the Collection.

collection - The Collection to get the children of.

Errors

If an error occurs when interacting with the storage system, or a child Element cannot be found, an error will be returned.

pub fn child_info_for<C: Collection>( parent_id: Id, collection: &C, ) -> Result<Vec<ChildInfo>, StorageError>

The basic info for children of the Collection.

This gets basic info for children of the Collection, which are the Elements that are directly below the Collection’s owner in the hierarchy. This is a simple method that returns the children as a list, and does not provide any filtering or ordering.

See children_of() for more information.

Parameters

parent_id - The ID of the parent entity that owns the Collection.

collection - The Collection to get the children of.

Errors

If an error occurs when interacting with the storage system, or a child Element cannot be found, an error will be returned.

See also

children_of()

pub fn compare_trees( foreign_entity_data: Option<Vec<u8>>, foreign_index_data: ComparisonData, ) -> Result<(Vec<Action>, Vec<Action>), StorageError>

Compares a foreign entity with a local one.

This function compares a foreign entity, usually from a remote node, with the version available in the tree in local storage, if present, and generates a list of Actions to perform on the local tree, the foreign tree, or both, to bring the two trees into sync.

The tuple returned is composed of two lists of actions: the first list contains the actions to be performed on the local tree, and the second list contains the actions to be performed on the foreign tree.

Parameters

foreign_entity - The foreign entity to compare against the local version. This will usually be from a remote node.

Errors

This function will return an error if there are issues accessing local data or if there are problems during the comparison process.

pub fn compare_affective( data: Option<Vec<u8>>, comparison_data: ComparisonData, ) -> Result<(), StorageError>

Compares a foreign entity with a local one, and applies the resulting actions to bring the two entities into sync.

Errors

This function will return an error if there are issues accessing local data or if there are problems during the comparison process.

pub fn find_by_id<D: Data>(id: Id) -> Result<Option<D>, StorageError>

Finds an Element by its unique identifier.

This will always retrieve a single Element, if it exists, regardless of where it may be in the hierarchy, or what state it may be in.

Parameters

id - The unique identifier of the Element to find.

Errors

If an error occurs when interacting with the storage system, an error will be returned.

pub fn find_by_id_raw(id: Id) -> Option<Vec<u8>>

Finds an Element by its unique identifier without deserialising it.

This will always retrieve a single Element, if it exists, regardless of where it may be in the hierarchy, or what state it may be in.

Notably it returns the raw bytes without attempting to deserialise them into a Data type.

Parameters

id - The unique identifier of the Element to find.

pub fn find_by_path<D: Data>(_path: &Path) -> Result<Vec<D>, StorageError>

Finds one or more Elements by path in the hierarchy.

This will retrieve all Elements that exist at the specified path in the hierarchy. This may be a single item, or multiple items if there are multiple Elements at the same path.

Parameters

path - The path to the Elements to find.

Errors

If an error occurs when interacting with the storage system, an error will be returned.

pub fn find_children_by_id<D: Data>( parent_id: Id, collection: &str, ) -> Result<Vec<D>, StorageError>

Finds the children of an Element by its unique identifier.

This will retrieve all Elements that are children of the specified Element. This may be a single item, or multiple items if there are multiple children. Notably, it will return None if the Element in question does not exist.

Parameters

parent_id - The unique identifier of the Element to find the children of.

collection - The name of the Collection to find the children of.

Errors

If an error occurs when interacting with the storage system, an error will be returned.

pub fn generate_comparison_data( id: Option<Id>, ) -> Result<ComparisonData, StorageError>

Generates comparison data for an entity.

This function generates comparison data for the specified entity, which includes the entity’s own hash, the full hash of the entity and its children, and the IDs and hashes of the children themselves.

Parameters

entity - The entity to generate comparison data for.

Errors

If an error occurs when interacting with the storage system, an error will be returned.

pub fn has_children<C: Collection>( parent_id: Id, collection: &C, ) -> Result<bool, StorageError>

Whether the Collection has children.

This checks whether the Collection of the specified entity has children, which are the entities that are directly below the entity’s Collection in the hierarchy.

Parameters

parent_id - The unique identifier of the entity to check for children.

collection - The Collection to check for children.

Errors

If an error occurs when interacting with the storage system, an error will be returned.

pub fn parent_of<D: Data>(child_id: Id) -> Result<Option<D>, StorageError>

Retrieves the parent entity of a given entity.

Parameters

child_id - The Id of the entity whose parent is to be retrieved.

Errors

If an error occurs when interacting with the storage system, an error will be returned.

pub fn remove_child_from<C: Collection>( parent_id: Id, collection: &C, child_id: Id, ) -> Result<bool, StorageError>

Removes a child from a collection.

Parameters

parent_id - The ID of the parent entity that owns the Collection.

collection - The collection from which the child should be removed.

child_id - The ID of the child entity to remove.

Errors

If an error occurs when interacting with the storage system, an error will be returned.

pub fn root<D: Data>() -> Result<Option<D>, StorageError>

Retrieves the root entity for a given context.

Parameters

context_id - An identifier for the context whose root is to be retrieved.

Errors

If an error occurs when interacting with the storage system, an error will be returned.

pub fn commit_root<D: Data>(root: Option<D>) -> Result<(), StorageError>

Saves the root entity to the storage system, and commits any recorded actions or comparisons.

This function must only be called once otherwise it will panic.

Errors

This function will return an error if there are issues accessing local data or if there are problems during the comparison process.

pub fn save<D: Data>(entity: &mut D) -> Result<bool, StorageError>

Saves an Element to the storage system.

This will save the provided Element to the storage system. If the record already exists, it will be updated with the new data. If the record does not exist, it will be created.

Update guard

If the provided Element is older than the existing record, the update will be ignored, and the existing record will be kept. The Boolean return value indicates whether the record was saved or not; a value of false indicates that the record was not saved due to this guard check — any other reason will be due to an error, and returned as such.

Dirty flag

Note, if the Element is not marked as dirty, it will not be saved, but true will be returned. In this case, the record is considered to be up-to-date and does not need saving, and so the save operation is effectively a no-op. If necessary, this can be checked before calling save()](crate::entities::Element::save()) by calling [is_dirty().

Hierarchy

It’s important to be aware of the hierarchy when saving an entity. If the entity is a child, it must have a parent, and the parent must exist in the storage system. If the parent does not exist, the child will not be saved, and an error will be returned.

If the entity is a root, it will be saved as a root, and the parent ID will be set to None.

When creating a new child entity, it’s important to call add_child_to() in order to create and associate the child with the parent. Thereafter, save() can be called to save updates to the child entity.

Merkle hash

The Merkle hash of the Element is calculated before saving, and stored in the Element itself. This is used to determine whether the data of the Element or its children has changed, and is used to validate the stored data.

Note that if the Element does not need saving, or cannot be saved, then the Merkle hash will not be updated. This way the hash only ever represents the state of the data that is actually stored.

Parameters

element - The Element whose data should be saved. This will be serialised and stored in the storage system.

Errors

If an error occurs when serialising data or interacting with the storage system, an error will be returned.

pub fn save_raw( id: Id, data: Vec<u8>, metadata: Metadata, ) -> Result<Option<[u8; 32]>, StorageError>

Saves raw data to the storage system.

Errors

If an error occurs when serialising data or interacting with the storage system, an error will be returned.

pub fn validate() -> Result<(), StorageError>

Validates the stored state.

This will validate the stored state of the storage system, i.e. the data that has been saved to the storage system, ensuring that it is correct and consistent. This is done by calculating Merkle hashes of the stored data, and comparing them to the expected hashes.

Errors

If an error occurs when interacting with the storage system, an error will be returned.