calimero_storage::store
Trait StorageAdaptor

pub trait StorageAdaptor {
    // Required methods
    fn storage_read(key: Key) -> Option<Vec<u8>>;
    fn storage_remove(key: Key) -> bool;
    fn storage_write(key: Key, value: &[u8]) -> bool;
}

Determines where the ultimate storage system is located.

This trait is mainly used to allow for a different storage location to be used for key operations during testing, such as modelling a foreign node’s data store.

Required Methods

fn storage_read(key: Key) -> Option<Vec<u8>>

Reads data from persistent storage.

Parameters

key - The key to read data from.

fn storage_remove(key: Key) -> bool

Removes data from persistent storage.

Parameters

key - The key to remove.

fn storage_write(key: Key, value: &[u8]) -> bool

Writes data to persistent storage.

Parameters

key - The key to write data to.

value - The data to write.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.